Simplify bi/trigram accounting
Ignore bi/trigram's containing whitespace. The rationale here is that
whitespace is a "reset", defining a word boundary which gives the
fingers time to return to their base position. This should make the
analysis simpler.

Likewise, ignoring case will make things simpler (ie "E" and "e" are
considered the same; we assume the opposite hand will take care of the
shifting (not necessarily a valid assumption, but definitely a
simplifying one).