Show "summary" stats for `layout-stats` subcommand
This is really a debugging tool, more than anything, to help me debug
the various weightings of the different effort multipliers. On the
sample corpus, I get this for Qwerty:

  getRollMultiplier: 0.8475
  getRowJumpMultiplier: 1.1045
  getSameFingerMultiplier: 1.0585
  getPositionMultiplier: 1.7837
  getFingerMultiplier: 0.5542

And this for Colemak:

  getRollMultiplier: 0.7920
  getRowJumpMultiplier: 1.0352
  getSameFingerMultiplier: 1.0279
  getPositionMultiplier: 1.3047
  getFingerMultiplier: 0.5474

Suggesting that the roll multiplier may be wrong (I would expect Colemak
to fare better), and the others may need some tweaking.