Attempt fix at resource leak
Running with a large value for `-r` we eventually run out of memory. I
first tried to fix this with a `clear()` call on the map, but that still
left some uncollected resources lying around. Reassigning the map so
that the whole thing can be GC'd is my second attempt; it seems to run
for longer in this case but still dies eventually. Next step will be to
do some actual profiling to see which objects are hanging around.